+++ packages/pizza_repository/lib/src/models/pizza.dart
@@ -8,8 +8,8 @@ class Pizza {
   int spicy;
   String name;
   String description;
-  double price;
-  double discount;
+  int price;
+  int discount;
   Macros macros;
 
   Pizza({
